requirements.txt
PyMuPDF
frontend
langchain
python-dotenv
matplotlib
pandas
numpy
sentence_transformers
scikit-learn
chromadb
transformers
accelerate
bitsandbytes
python-dotenv
hf_xet
streamlit